gcs_trigger_cloud_run/cloudrun/requirements.txt
uvicorn==0.29.0
fastapi[all]==0.111.0
google-cloud-storage==2.16.0
cloudevents==1.10.1
